py/pyproject.toml
[project]
name = "py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "click>=8.1.8",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "requests-cache>=1.2.1",
]

[dependency-groups]
dev = [
    "ruff>=0.11.1",
]
